# api/src/main/resources/openapi/expense-api.yaml
openapi: 3.0.0
info:
  title: Expense Management API
  description: RESTful API for managing expenses, categories, and generating reports
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /api/expenses:
    post:
      operationId: createExpense
      summary: Create a new expense
      tags:
        - Expenses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseApiRequest'
      responses:
        '201':
          description: Expense created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseApiResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
    
    get:
      operationId: getAllExpenses
      summary: Retrieve all expenses with optional filters
      tags:
        - Expenses
      parameters:
        - name: startDate
          in: query
          description: Filter expenses from this date (inclusive)
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter expenses up to this date (inclusive)
          required: false
          schema:
            type: string
            format: date
        - name: categories
          in: query
          description: Filter by one or more categories
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: status
          in: query
          description: Filter by expense status
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExpenseApiResponse'
        '400':
          description: Invalid filter parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
  
  /api/expenses/{id}:
    get:
      operationId: getExpenseById
      summary: Retrieve a specific expense by ID
      tags:
        - Expenses
      parameters:
        - name: id
          in: path
          description: Expense ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Expense details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseApiResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
    
    put:
      operationId: updateExpense
      summary: Update an existing expense
      tags:
        - Expenses
      parameters:
        - name: id
          in: path
          description: Expense ID
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseApiRequest'
      responses:
        '200':
          description: Expense updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseApiResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
    
    delete:
      operationId: deleteExpense
      summary: Delete an expense
      tags:
        - Expenses
      parameters:
        - name: id
          in: path
          description: Expense ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Expense deleted successfully
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'
  
  /api/categories:
    get:
      operationId: getAllCategories
      summary: Retrieve all available expense categories
      tags:
        - Categories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
  
  /api/reports:
    get:
      operationId: generateReport
      summary: Generate expense report with aggregated statistics
      tags:
        - Reports
      parameters:
        - name: startDate
          in: query
          description: Filter expenses from this date (inclusive)
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: Filter expenses up to this date (inclusive)
          required: false
          schema:
            type: string
            format: date
        - name: categories
          in: query
          description: Filter by one or more categories
          required: false
          schema:
            type: array
            items:
              type: string
          style: form
          explode: true
        - name: status
          in: query
          description: Filter by expense status
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Expense report with aggregated data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseReportApiResponse'
        '400':
          description: Invalid filter parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorApiResponse'

components:
  schemas:
    ExpenseApiRequest:
      type: object
      required:
        - amount
        - date
        - category
        - description
      properties:
        amount:
          type: number
          format: double
          description: Expense amount (must be positive)
          example: 125.50
        date:
          type: string
          format: date
          description: Expense date (cannot be in the future)
          example: "2024-01-15"
        category:
          type: string
          description: Expense category
          example: "TRAVEL"
        description:
          type: string
          description: Expense description
          maxLength: 500
          example: "Flight to New York"
        status:
          type: string
          description: Expense status (optional for updates)
          example: "PENDING"
    
    ExpenseApiResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique expense identifier
          example: 1
        amount:
          type: number
          format: double
          description: Expense amount
          example: 125.50
        date:
          type: string
          format: date
          description: Expense date
          example: "2024-01-15"
        category:
          type: string
          description: Expense category
          example: "TRAVEL"
        description:
          type: string
          description: Expense description
          example: "Flight to New York"
        status:
          type: string
          description: Expense status
          example: "PENDING"
        createdAt:
          type: string
          format: date-time
          description: Timestamp when expense was created
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when expense was last updated
          example: "2024-01-15T10:30:00Z"
    
    ExpenseReportApiResponse:
      type: object
      properties:
        totalAmount:
          type: number
          format: double
          description: Sum of all expense amounts
          example: 1250.75
        expenseCount:
          type: integer
          format: int64
          description: Total number of expenses
          example: 10
        categorySubtotals:
          type: object
          additionalProperties:
            type: number
            format: double
          description: Subtotals grouped by category
          example:
            TRAVEL: 500.00
            FOOD: 250.50
            OFFICE_SUPPLIES: 500.25
        startDate:
          type: string
          format: date
          description: Start date of the report period
          example: "2024-01-01"
        endDate:
          type: string
          format: date
          description: End date of the report period
          example: "2024-01-31"
        generatedAt:
          type: string
          format: date-time
          description: Timestamp when report was generated
          example: "2024-01-15T10:30:00Z"
    
    ErrorApiResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: Error code identifier
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Validation failed for expense creation"
        details:
          type: array
          items:
            type: string
          description: List of detailed error messages
          example:
            - "Amount must be positive"
            - "Date cannot be in the future"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when error occurred
          example: "2024-01-15T10:30:00Z"